requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==2.2.0
django-import-export==2.0.1
django-jsonfield==1.4.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
gunicorn==20.0.4
importlib-metadata==1.7.0
jdcal==1.4.1
jsonfield==3.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
six==1.11.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
tablib==0.14.0
whitenoise==3.3.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
ipdb==0.10
